+++ src/main/java/hotspot/user/policy/controller/response/AppliedPolicyResponse.java
@@ -12,7 +12,7 @@ public record AppliedPolicyResponse(
         Long memberId,
         String memberName,
         Long subId,
-        int dataLimit, // 개인별 한도
+        double dataLimit, // 개인별 한도
         int priority, // 구성원 내 우선 순위
         List<BlockPolicyResponse> blockPolicyResponseList,
         List<AppBlockedServiceResponse> appBlockedServiceResponseList
